Lesson2.Bugs fixing

+++ product/src/main/java/ru/geekbrains/product/ProductService.java
@@ -25,7 +25,7 @@ public class ProductService {
         return sum / size;
     }
 
-    public float getSize() {
+    public int getSize() {
         return productContainer.getProducts().size();
     }
